W182: the wide-local sweep -- roughly 130 functions retyped, two MATCHes, four more windows cleared.
The defect is small and the blast radius was enormous. A 64-bit local holding a
32-bit call result makes MWCC emit a spurious dsll32/dsra32 sign-extension pair
at every use, and the raw decompiler produced these by the hundred. Eight
agents worked the 286-function census in `build/wip/W182_widelocal.md` and
narrowed the local -- and only the local, one call site at a time, measured --
wherever the retail bytes showed no extension belonged there.

MATCHes:
  y_runtime 00431110  nd9 -> 0   the hand-expanded signed divide-by-two
                                 temporary replaced by `param_3 / 2`.
  k_clump   001a7570  nd2 -> 0   the real condition recovered: state[1] must be
                                 zero AND either state[0] is zero or every
                                 remaining list is zero. Retail's beqz
                                 orientation falls out of the correct predicate;
                                 the branch shape was never the defect.

Windows cleared: y_runtime 00448060, 00448cf0 and 0045a490, and fclCombine
003d3d20. Over-window count 139 -> 135.

Total normalized-diff reduction across the tree: 3,739. The largest single
movers were not the big functions but the mid-size ones where one wrong width
poisoned a dozen uses:
  mdlFile   0034f910  nd361 -> 227  the exact 23-instruction VU conversion
                                    reinstated, state and loop-counter
                                    lifetimes separated, and a typed u32
                                    implementation alias that removed both the
                                    spurious 64-bit parameter extension and the
                                    0x110 frame. The frame is now retail's
                                    0x100 and the earlier over-window
                                    regression is gone.
  mdlFile   0034d510  nd236 -> 113
  mdlFile   00356c70  nd513 -> 362
  mdlFile   00357550  nd221 -> 154
  mdlFile   0034cc00  nd505 -> 387  the loop-live typed 0.0f recovered, giving
                                    retail's separate f20 zero lifetime beside
                                    f21 scale.
  mt_scene  003b7090  nd529 -> 387
  fclShopMisc 0040a6d0 nd132 -> 19, 0040b910 nd1284 -> 1200, and thirty more
  y_runtime 00448060  nd2658 -> 2428, also replacing CONCAT44 and manual
                                    64-bit sign comparisons with direct signed
                                    char comparisons
  itfPanel  003ad1b0  nd126 -> 95, 003ada00's real width defect removed via a
                                    u32 pragma alias used at ONE call site so
                                    the sibling u64 call stayed intact
  mt_evtMain 003655f0 nd284 -> 214, 0036cc30 nd326 -> 260

The negative half of the sweep is equally recorded. Roughly twenty narrowings
were measured and REVERTED because they grew the object or worsened the diff:
btlFormation 002bb0b0 improved nd322 -> 219 but crossed its window; mdlFile
0031d9c0, 0031ebe0, 0031ef80 and 00322010 all regressed; y_runtime 00438010 and
0043b980 grew. Several census hits turned out to be genuine 64-bit values --
packed Camp coordinates, a real aggregate copy in mdlFile 00322010, y_runtime's
retained unsigned conversion expansions -- and were left alone. The census
finds candidates; only the retail bytes decide.

Two structural notes for the next pass. mdlFile 003243f0's dsll32/dsra32 comes
from the function's own u64 parameter while the callee already expects u32, but
the fix needs a forward declaration outside the agent's assigned address block,
so it was correctly left undone rather than reached for. And mdlFile 00335180 /
003377f0 gained back two hidden arguments to FUN_00323860 that retail really
passes -- byte-neutral, but the call semantics are now truthful.

Checkpoint: 6423 / 8211 first-party MATCH (78.2%). Zero lost, zero new
over-window addresses, zero compile errors, zero stale markers.

+++ src/Camp/_h_camp_system.c
@@ -453,10 +453,10 @@ extern void FUN_001159F0(f32 x, f32 y, u32 drawContext);
 extern void FUN_00115BC0(f32 x, f32 y, u32 drawContext);
 extern void FUN_00115AD0(f32 x, f32 y, u32 drawContext);
 extern s32 FUN_0017D2E0(u16 arg0);
-extern u64 FUN_0030BB40(u16 arg0);
-extern u32 FUN_00523AC8(void *dst, u32 stringId, u64 arg2);
+extern u32 FUN_0030BB40(u16 arg0);
+extern u32 FUN_00523AC8(void *dst, u32 stringId, u32 arg2);
 #pragma alias campCE50FormatText FUN_00523AC8
-extern void campCE50FormatText(void *dst, const char *format, u64 arg);
+extern void campCE50FormatText(void *dst, const char *format, u32 arg);
 extern void FUN_003B32D0(u32 drawContext, s32 x, s32 y, u32 color, s32 font, s32 flags, void *text, s32 width, s32 height);
 #pragma alias campCE50DrawSprite FUN_001159F0
 #pragma alias campSystemDrawSprite FUN_001159F0
@@ -502,7 +502,7 @@ extern u64 FUN_0016C970(u16 arg0);
 extern s16 FUN_0016C920(u16 arg0);
 extern u32 FUN_0016C670(u16 arg0);
 extern u32 FUN_0016C5F0(u16 arg0);
-extern u64 FUN_00177790(u16 arg0);
+extern u32 FUN_00177790(u16 arg0);
 
 void *FUN_0015B430(KwlnTask *task);
 void FUN_0015C520(void *list, s32 ownerMode);
@@ -1150,7 +1150,7 @@ void FUN_0015CE50(u64 packedPosition, void *listPointer, s32 alpha,
     s32 i;
     s32 record_index;
     uintptr_t list;
-    u64 text_handle;
+    u32 text_handle;
     f32 x;
     f32 y;
     f32 row_x;
@@ -1857,7 +1857,7 @@ void FUN_0015FA90(CampDrawRecord *record, s32 recordIndex, CampDrawWork *work)
     s16 itemWord;
     s16 category;
     u8 alpha;
-    u64 spriteHandle;
+    u32 spriteHandle;
     CampD8Object *sprite;
     u8 *groupBase;
